adapted to moved SimuLizarWorkflowConfiguration
